% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_d.R
\docType{data}
\name{d}
\alias{d}
\title{Example data from HW 1}
\format{An object of class \code{data.frame} with 30 rows and 2 columns.}
\usage{
d
}
\description{
Example data from HW 1
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/q1p4.txt}
}
\author{
Charlie Geyer \email{geyer@umn.edu}
}
\keyword{datasets}
