% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{dplyr}
\alias{dplyr}
\title{wrapper function for dplyr for data farmData}
\usage{
dplyr(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
data.frame
}
\description{
wrapper function for dplyr for data farmData
}
\examples{
data(farmData)
dplyr(farmData)
}
