% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_farmData.R
\docType{data}
\name{farmData}
\alias{farmData}
\title{Example data from Lecture 7}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9597 rows and 1754 columns.}
\usage{
farmData
}
\description{
Example data from Lecture 7
}
\references{
\url{http://users.stat.umn.edu/~almquist/3811_examples/data.csv}
}
\author{
Zack Almquist \email{almquist@umn.edu}
}
\keyword{datasets}
