% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_Quiz_Functions.R
\name{func12}
\alias{func12}
\title{Matrix Standardization}
\usage{
func12(a)
}
\arguments{
\item{a}{matrix}
}
\value{
matrix
}
\description{
Takes a matrix and standardizes its columns
}
\details{
Quiz 2 Problem 3
}
\examples{
load(url("http://www.stat.umn.edu/geyer/3701/data/q2p3.rda"))
func12(a)
}
