% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_Quiz_Functions.R
\name{func13}
\alias{func13}
\title{Matrix Standardization (again)}
\usage{
func13(a)
}
\arguments{
\item{a}{matrix}
}
\value{
matrix
}
\description{
Takes a matrix and standardizes its columns
}
\details{
HW 2 Problem 4
}
\examples{
load(url("http://www.stat.umn.edu/geyer/3701/data/q2p3.rda"))
func13(a)
}
