% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW1_P7.R
\name{func4}
\alias{func4}
\title{Hw1 Problem 7}
\usage{
func4(x, func, interval)
}
\arguments{
\item{x}{data.frame}

\item{func}{function}

\item{interval}{object}
}
\value{
scalar
}
\description{
Hw1 Problem 7
}
\examples{
x1 <- scan(url("http://www.stat.umn.edu/geyer/3701/data/q1p3.txt"))
x2 <- scan(url("http://www.stat.umn.edu/geyer/3701/data/q1p7c.txt"))
x3 <- scan(url("http://www.stat.umn.edu/geyer/3701/data/q1p7b.txt"))
func1 <- function(theta,x) dgamma(x, shape = theta, log = TRUE)
func2 <- function(theta,x) dcauchy(x, location = theta, log = TRUE)
func3 <- function(theta,x) dbinom(x, 20, prob = 1/(1+exp(-theta)), log = TRUE)
result_gamma <- func4(x1, func1, c(0,3))
result_cauchy <- func4(x2, func2, c(0,100))
result_dbinom <- func4(x3, func3, c(-10,10))  

}
