% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_Quiz_Functions.R
\name{func5}
\alias{func5}
\title{Calculate Mean, Variane, SD}
\usage{
func5(x)
}
\arguments{
\item{x}{vector}
}
\value{
list
}
\description{
Computes the mean, variance and sd of a vector
}
\details{
Quiz 1 Problem 1
}
\examples{
func5(rnorm(10))
}
