% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_Quiz_Functions.R
\name{func6}
\alias{func6}
\title{Calculate Mean, Variane, SD (again)}
\usage{
func6(x)
}
\arguments{
\item{x}{vector}
}
\value{
list
}
\description{
Computes the mean, variance and sd of a vector, but with user checks
}
\details{
Quiz 1 Problem 2
}
\examples{
func6(rnorm(10))
}
