% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_Quiz_Functions.R
\name{func7}
\alias{func7}
\title{MLE of gamma distribution}
\usage{
func7(x)
}
\arguments{
\item{x}{vector}
}
\value{
scalar
}
\description{
Computes the liklihood of a gamma distribution
}
\details{
Quiz 1 Problem 3
}
\examples{
func7(rnorm(10))
}
