% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_Quiz_Functions.R
\name{func8}
\alias{func8}
\title{Weighted mean, var, sd}
\usage{
func8(d)
}
\arguments{
\item{d}{data.frame}
}
\value{
list
}
\description{
Computes the weighted mean, var, sd
}
\details{
HW 1 Problem 4
}
\examples{
data(d)
func8(d)
}
