% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_Quiz_Functions.R
\name{func9}
\alias{func9}
\title{Highlevel check function}
\usage{
func9(x)
}
\arguments{
\item{x}{object}
}
\value{
object
}
\description{
Checks and throws error if not numeric, finit, zero lenth, NA, NAN
}
\details{
HW1 Problem 6
}
\examples{
func9(NA)
}
