% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_Quiz_Functions.R
\name{\%func11\%}
\alias{\%func11\%}
\title{Matrix Multiplication (again)}
\usage{
A \%func11\% x
}
\arguments{
\item{A}{matrix}

\item{x}{vector}
}
\value{
matrix
}
\description{
Takes a matrix A and a vector x and calculates xT*A^-1*x
}
\details{
Quiz 2 Problem 2
}
\examples{
load(url("http://www.stat.umn.edu/geyer/3701/data/q2p1.rda"))
a \%func11\% x
}
