% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_growth.R
\docType{data}
\name{growth}
\alias{growth}
\title{Example data from data}
\format{An object of class \code{data.frame} with 93 rows and 33 columns.}
\usage{
growth
}
\description{
Example data from data
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/growth.csv}
}
\author{
Charlie Geyer \email{geyer@umn.edu}
}
\keyword{datasets}
