% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_Quiz_Functions.R
\name{myapply}
\alias{myapply}
\title{Handmade Apply Function}
\usage{
myapply(X, MARGIN, FUN, ...)
}
\arguments{
\item{X}{array}

\item{MARGIN}{object}

\item{FUN}{function}
}
\value{
array
}
\description{
Functions the same as the apply() function
}
\details{
HW2 Problem 5
}
\examples{
fred <- matrix(1:6, ncol = 2)
myapply(fred,1,"mean")
myapply(fred, 2, "mean")
myapply(fred, 1, "max")
myapply(fred, 2, "max")
myapply(fred, 1, function(x) quantile(x,0.75))
myapply(fred, 2, function(x) quantile(x,0.75))
myapply(fred,1,"quantile",probs=.75)
myapply(fred,2,"quantile",probs=.75)
myapply(fred,1,"quantile",probs=(1:3)/4)
myapply(fred,2,"quantile",probs=(1:3)/4)
}
