% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdata.R
\name{plotmydata}
\alias{plotmydata}
\title{Wrapper function for ggplot2 for data d}
\usage{
plotmydata(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
ggplot2
}
\description{
Wrapper function for ggplot2 for data d
}
\examples{
data(farmData)
plotmydata(farmData)
}
